+++ Game/CubeInterface.h
@@ -15,7 +15,7 @@ class CubeInterface
     void clearLayer(byte layer);
     void clearColumn(byte x, byte y);
     void writeCube();
-    void wait(byte t);
+    void wait(int t);
     boolean ledStatus[8][8][8];
   private:
     void highBit();
